watch-and-learn-main/test_cases/competitor_analysis/social_tracker/requirements.txt
# LangGraph Social Media Tracker Dependencies

# Core LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Web Scraping
firecrawl-py>=1.0.0

# Data Processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
httpx>=0.27.0
aiohttp>=3.9.0
